added buying balance, base dropdown, user info,

+++ feature/reservation/data/src/main/java/ge/tbca/city_park/reservation/data/model/ReservationDTO.kt
@@ -12,6 +12,6 @@ data class ReservationDTO(
     val createdAt: Long,
     val active: Boolean = true,
     val endedAt: Long? = null,
-    val cost: Long? = null,
+    val cost: Double? = null,
     val duration: Long? = null
 )
