+++ feature/reservation/presentation/src/main/java/ge/tba/city_park/reservation/presentation/model/ReservationUi.kt
@@ -8,5 +8,5 @@ data class ReservationUi(
     val createdAt: String,
     val active: Boolean = true,
     val endedAt: String? = null,
-    val cost: Long? = null
+    val cost: Double? = null
 )
